services/discord/chat/requirements.txt
discord
pydantic-settings
fastapi[standard]
structlog
opentelemetry-api
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-aiohttp-client
opentelemetry-sdk